+++ modules/client/CHANGELOG.md
@@ -13,6 +13,16 @@ TEMPLATE:
 
 -->
 ## [UPCOMING]
+## [1.7.0]
+### Added
+- Add `delegateTokens` function in `TokenVotingClient`
+- Add `undelegateTokens` function in `TokenVotingClient`
+- Add `getDelegatee` function in `TokenVotingClient`
+- Add `prepareUninstallation` function in `Client`
+### Changed
+- Breaking: `getMembers` function in `TokenVotingClient` no loger returns an `string[]`. Now it returns `TokenVotingMember[]` which includes the delegation status of each member
+- `condition` is now an optional parameter in `MultiTargetPermission`  
+## [1.6.0]
 ### Fixed
 - Fixed plugin version on the dao subgraph queries
 ### Added
@@ -21,8 +31,6 @@ TEMPLATE:
 
 ### Changed
 - Properties `signer` and `web3providers` are now optional in the `ContextParams`
-
-
 ## [1.5.0]
 ### Added
 - Add `wrapTokens` and `unwrapTokens` functions to support token contract without support for snapshots